% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci2se}
\alias{ci2se}
\title{Convert normal-based confidence intervals into standard errors}
\usage{
ci2se(ci, trans = identity, level = 0.95)
}
\arguments{
\item{ci}{a 2-row matrix of confidence intervals, with lower limit
in the first row and upper limit in the second row}

\item{trans}{a function that transforms the confidence intervals
into normal space.}

\item{level}{the level of the confidence intervals supplied}
}
\value{
a vector of standard errors
}
\description{
This function returns the standard errors that would give a
normally-based confidence interval of given width with the
specified level.
}
\examples{
inst/examples/eg_risk_relative.R
}
