% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format.perc}
\alias{format.perc}
\title{Format a decimal as a percentage}
\usage{
\method{format}{perc}(probs, digits)
}
\arguments{
\item{probs}{vector of probabilities}

\item{digits}{number of digits to display}
}
\description{
Copied shamelessly from \code{stats:::format.perc}.
}
\keyword{internal}
