% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat2arr}
\alias{mat2arr}
\title{Convert a matrix into a 3d array in order by factors}
\usage{
mat2arr(data, rowf, colf)
}
\arguments{
\item{data}{the matrix}

\item{rowf, colf}{factors of length \code{nrow(data)}}
}
\description{
This has undefined behaviour when
\code{anyDuplicated(interaction(rowf,colf))}, which is not tested.
}
\details{
We want \code{rowf} for dim 1, \code{colf} for dim 2, and the
columns of \code{data} for dim 3.
}
