% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match.data.frame}
\alias{match.data.frame}
\title{Value matching in data frames}
\usage{
match.data.frame(x, table, by = intersect(names(x), names(table)),
  nomatch = NA_integer_, incomparables = NULL)
}
\arguments{
\item{x}{the data frame to be matched}

\item{table}{reference rows}

\item{by}{columns of interest in each data frame}

\item{nomatch, incomparables}{see \code{\link{match}}}
}
\value{
an integer vector of length \code{nrow(x)}
}
\description{
Find row indices of data frame that match a reference
}
