% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{ModifyCategorical}
\alias{ModifyCategorical}
\title{Simple to/from dataframe modification function}
\usage{
ModifyCategorical(factor, to, from = NULL, ...)
}
\arguments{
\item{factor}{the name of the factor column to be changed}

\item{to}{the target level of the factor}

\item{from, ...}{the source levels of the factor (defaulting to all levels if
unspecified)}
}
\value{
a function that takes a data frame and outputs a modified data frame
}
\description{
Returns a function that changes all \code{from} levels of a column in a data
frame to the \code{to} level, and returns the modified data frame. If
\code{from} is not specified, then the entire column is replaced.
}
\seealso{
\code{\link{paf}} and \code{\link{paf_ci}}, where this function can
  be used
}
