% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf.R
\name{paf_ext}
\alias{paf_ext}
\alias{paf_ext_ci}
\title{PAF estimates with external prevalences}
\usage{
paf_ext(risk, df, prevalence, mod_prev, ...)

paf_ext_ci(risk, df, prevalence, mod_prev, var_prev, ..., se_fit = TRUE,
  method = c("delta"), level = 0.95)
}
\arguments{
\item{risk}{a \code{\link{risk-object}}}

\item{df}{a data frame describing the new population}

\item{prevalence}{a vector of external prevalence estimates corresponding to
rows in \code{df}}

\item{mod_prev}{a transition matrix to modify \code{df}; see Details}

\item{...}{further arguments passed to \code{risk$riskfn} and
\code{risk$dtransvar}}

\item{var_prev}{variance-covariance matrix for the prevalence estimates}

\item{se_fit}{if \code{TRUE}, then return the estimated standard error in
transformed space}

\item{method}{method to calculate confidence intervals (currently the delta
method alone is supported)}

\item{level}{the confidence level required, default 0.95}
}
\description{
Calculates a point estimate (\code{paf_ext}) or a point estimate with
confidence interval (\code{paf_ext_ci}) for a PAF given an population and
prevalence estimates, information about the risk and a counterfactual
distribution.
}
\details{
If \code{df} is missing and the original population (used for risk
estimation) can be accessed in \code{risk}, then the PAF will be calculated
from the original population.

In this function, the counterfactual population is defined using a transition
matrix \eqn{A} = \code{mod_prev} such that, if \eqn{p} = \code{prevalence} is
a vector of prevalences, then the counterfactual prevalence is \eqn{Ap}. Note
that the columns of \code{mod_prev} should all sum to 1, and that
\code{prevalence} is reweighted to sum to one.
}
\seealso{
\code{\link{paf}} for PAF calculations with explicit populations;
  \code{\link{TransitionSimple}} and related functions to generate transition
  matrices.
}
