% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk-object}
\alias{risk-object}
\title{Risk function information}
\description{
A standardised object containing information regarding risk
functions needed to calculate PAFs.
}
\details{
A risk object is a list that contains the following elements:

\describe{

\item{\code{riskfn(df, ...)}}{A function which takes a population
as described in \code{df} and optional further arguments, and
outputs the event risk for each member in the population.}

\item{\code{dtransvar(df, ...)}}{A function that returns the
derivatives of risks with respect to any underlying random
parameters. The output should be a matrix with element in row
\eqn{i} and column \eqn{j} \deqn{J_{i,j} = \frac{\partial r(x_i;
\beta)}{\partial \beta_j}} where \eqn{x_i} corresponds to row
\eqn{i} of the input dataframe.}

\item{\code{var(d)}}{A function that for any vector \code{d},
returns the variance of the dot product \eqn{d^\top \beta}.}

\item{\code{source_df}}{(optional) The source population used for
risk function calculation.}

}
}
\seealso{
Other risk functions: \code{\link{risk_glm}}
}
\concept{risk functions}
