% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-competing.R
\name{risk_cr}
\alias{risk_cr}
\title{Risk information for constant piecewise hazards model with
competing risks}
\usage{
risk_cr(resp1, resp2, predictors, data, breaks, ...)
}
\arguments{
\item{data}{the data frame from which observations are taken}

\item{breaks}{a vector of time points demarcating episodes of
constant hazard, *including* beginning and end points}

\item{...}{further arguments to be passed to
\code{\link[survival]{survreg}}}

\item{formula}{the model formula, which may include \code{.period}
(see Details)}
}
\value{
a \code{\link{risk-object}} list
}
\description{
This function returns a \code{\link{risk-object}} based on two
parametric piecewise constant hazards models, one each for the
primary and secondary events.
}
\details{
This function creates time-varying covariates from \code{data}, and
for that reason, the column names \code{.period}, \code{.ID},
\code{.tstart}, \code{.tstop} and \code{.event} should *not* be
used. These column names can be used in \code{formula} however,
and any arguments specified in \code{...} passed on to
\code{\link[survival]{survreg}}, and are taken to mean \describe{

\item{\code{.period}}{the time period varying as a factor}

\item{\code{.ID}}{a unique identifier for each input row}

\item{\code{.tstart}, \code{.tstop}}{start and end times of each
observation period, often the endpoints of the time period (if
censoring occurs)}

\item{\code{.event}}{an event indicator}

}

Note that the response in \code{formula} should be specified
without any reference to these intermediate variables.
}
\seealso{
Other risk functions: \code{\link{risk-object}},
  \code{\link{risk_glm}}, \code{\link{risk_pch}},
  \code{\link{risk_relative}}, \code{\link{risk_survreg}}
}
\concept{risk functions}
