% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_glm}
\alias{risk_glm}
\title{Risk information for binary GLMs}
\usage{
risk_glm(model)
}
\arguments{
\item{model}{a fitted GLM}
}
\value{
a \code{\link{risk-object}} list
}
\description{
This function extracts information on the risk from generalised linear models
with a binary predictor. This can be used for logistic regression, or other
models with binary dependent variables, such as the probit model.
}
\seealso{
Other risk functions: \code{\link{risk-object}},
  \code{\link{risk_cr}}, \code{\link{risk_pch}},
  \code{\link{risk_relative}}, \code{\link{risk_survreg}}
}
\concept{risk functions}
