% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_manual}
\alias{risk_manual}
\title{Manual risk input}
\usage{
risk_manual(table, risk, Jref, varmat)
}
\arguments{
\item{table}{reference risk factor data frame}

\item{risk}{risks corresponding to rows in \code{table}}

\item{Jref}{reference Jacobian matrix with rows corresponding to \code{table}}

\item{varmat}{variance matrix of underlying coefficients}
}
\value{
a \code{\link{risk-object}} list
}
\description{
Basic risk object function that takes risk estimates, gradients with respect
to coefficients, and coefficient variances.
}
\seealso{
Other risk functions: \code{\link{risk-object}},
  \code{\link{risk_cr}}, \code{\link{risk_glm}},
  \code{\link{risk_pch}}, \code{\link{risk_relative}},
  \code{\link{risk_survreg}}
}
\concept{risk functions}
