% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_relative}
\alias{risk_relative}
\title{Risk information from relative risks and their standard errors}
\usage{
risk_relative(RR, se.trans, eta = 1)
}
\arguments{
\item{RR}{a named list of relative risk estimates; see Examples}

\item{se.trans}{a named list of standard error estimates for log-relative risks}

\item{eta}{number between 0 and 1, ranging from least to most conservative}
}
\value{
a \code{\link{risk-object}} list
}
\description{
This function synthesises relative risk estimates and standard
errors from a specified set of factors. The covariance matrix is
estimated conservatively according to the \code{eta} parameter:
\eqn{\eta = 1} corresponds to the "worst-case" variance of the PAF,
while \eqn{\eta = 0} corresponds to uncorrelated log-relative
risks.
}
\examples{
# See Laaksonen et al. (2011) for these risk estimates
RR <- list(
  SEX = c("Male" = 1, "Female" = 1.0302),
  SMOKE = c("Never" = 1, "Former" = 1.4179, "<30/day" = 1.1446, ">=30/day" = 2.7868)
)

ci <- list(
  SEX = rbind(
    "lwr" = c("Male" = 1, "Female" = 0.7504),
    "upr" = c("Male" = 1, "Female" = 1.4144)
  ),
  SMOKE = rbind(
    "lwr" = c("Never" = 1, "Former" = 0.9863, "<30/day" = 0.7724, ">=30/day" = 1.2521),
    "upr" = c("Never" = 1, "Former" = 2.0385, "<30/day" = 1.6962, ">=30/day" = 6.2023)
  )
)

se <- lapply(ci, ci2se, trans = log)
se

risk <- risk_relative(RR, se, eta = 1)

paf_ci(risk, minifhs, ModifyCategorical("SMOKE", "Never", "<30/day", ">=30/day"), na.action = na.pass)
}
\seealso{
Other risk functions: \code{\link{risk-object}},
  \code{\link{risk_cr}}, \code{\link{risk_glm}},
  \code{\link{risk_manual}}, \code{\link{risk_pch}},
  \code{\link{risk_survreg}}
}
\concept{risk functions}
