% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_survreg}
\alias{risk_survreg}
\title{Risk information for location-scale survival models}
\usage{
risk_survreg(model)
}
\arguments{
\item{model}{a \code{survreg} object}
}
\value{
a \code{\link{risk-object}} list
}
\description{
This function extracts information on risk from survival models based on a
location-scale distribution.
}
\seealso{
Other risk functions: \code{\link{risk-object}},
  \code{\link{risk_cr}}, \code{\link{risk_glm}},
  \code{\link{risk_pch}}, \code{\link{risk_relative}}
}
\concept{risk functions}
