% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{TransitionComplete}
\alias{TransitionComplete}
\title{Simple all-to-one transition matrix}
\usage{
TransitionComplete(df, column, to, stay = NULL)
}
\arguments{
\item{df}{a data frame of risk factors, see Warning}

\item{column}{the name of the risk factor of interest}

\item{to}{the population target level in \code{column}}

\item{stay}{names of levels in \code{column} that should *not* be
modified}
}
\value{
a square transition matrix
}
\description{
Generates a transiton matrix where members of the population take
the specified level of a particular risk factor, with all other
risk factors unaffected.
}
\details{
Suppose that \code{df} is a data frame of risk factors, with risk
factor of interest (RFOI) in \code{df[[column]]}. We are interested
in the counterfactual scenario where every level of RFOI (except
those in \code{stay}) are converted to \code{to}.
}
\section{Warning}{


  If \code{df} contains any extraneous columns, then the dispersion
  among groups may not be accurate and this function has undefined
  behaviour.
}

\seealso{
Other transition functions: \code{\link{TransitionSimple}}
}
\concept{transition functions}
