% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{TransitionSimple}
\alias{TransitionSimple}
\title{Simple to/from transition matrix}
\usage{
TransitionSimple(df, column, from, to, prop = NULL)
}
\arguments{
\item{df}{a data frame of risk factors; see Warning}

\item{column}{the name of the risk factor of interest}

\item{from}{the source population level in \code{column}}

\item{to}{the target population level(s) in \code{column}}

\item{prop}{counterfactual dispersion proportions; see Details}
}
\value{
a square transition matrix
}
\description{
Generates a simple transition matrix where members with a particular risk
factor level are distributed amongst the other levels, holding all other risk
factors constant.
}
\details{
Suppose that \code{df} is a data frame of risk factors, with the risk factor
of interest (RFOI) in \code{df[[column]]}. We are interested in the
counterfactual scenario when those taking RFOI level \code{from} are replaced
with RFOI level \code{to}.

This function allows specification of multiple \code{to} levels. This can be
done by specifying \code{to} as a character vector.

By default, the current \code{from} population will be evenly dispersed among
the \code{to} population(s). This can be changed by specifying the
\code{prop} argument, which specifies the proportions of the \code{from}
population that move to the equivalent \code{to} populations. \code{prop}
should be a non-negative numeric vector of the same length as \code{to} that
sums to at most 1. If it sums to less than one, then the counterfactual
scenario will include the remainder in the original \code{from} population.
}
\section{Warning}{


  If \code{df} contains any extraneous columns, then the dispersion among
  groups may not be accurate and this function has undefined behaviour.
}

\seealso{
Other transition functions: \code{\link{TransitionComplete}}
}
\concept{transition functions}
