% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{VarMultinom}
\alias{VarMultinom}
\title{Variance matrix of multinomial distribution}
\usage{
VarMultinom(p, n)
}
\arguments{
\item{p}{A vector of prevalences which sums to 1, or an integer vector of
counts if \code{n} is missing.}

\item{n}{The size of the sample used to find \code{p}.}
}
\value{
a square covariance matrix
}
\description{
Estimates the variance-covariance matrix of a multinomial distribution.
}
