% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec2mat}
\alias{vec2mat}
\title{Rearrange vector to matrix in order by factors}
\usage{
vec2mat(data, rowf, colf)
}
\arguments{
\item{data}{vector of matrix entries}

\item{rowf, colf}{factors for row and column respectively of the
same length as data}
}
\value{
a matrix
}
\description{
This has undefined behaviour when
\code{anyDuplicated(interaction(rowf,colf))}, which is not tested.
}
\keyword{internal}
