% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf.R
\name{paf}
\alias{paf}
\alias{paf_ci}
\title{PAF estimates in a specific population}
\usage{
paf(risk, df, mod_df = identity, ...)

paf_ci(risk, df, mod_df = identity, ..., se_fit = TRUE,
  method = c("delta"), level = 0.95)
}
\arguments{
\item{risk}{a \code{\link{risk-object}}}

\item{df}{a data frame describing the new population}

\item{mod_df}{a function to modify \code{df}; see Details}

\item{...}{further arguments passed to \code{risk$riskfn} and
\code{risk$dtransvar}}

\item{se_fit}{if \code{TRUE}, then return the estimated standard error in
transformed space}

\item{method}{method to calculate confidence intervals (currently the delta
method alone is supported)}

\item{level}{the confidence level required, default 0.95}
}
\value{
the PAF estimate as a length-one numeric vector
}
\description{
Calculates a point estimate (\code{paf}) or a point estimate with confidence
interval (\code{paf_ci}) for a PAF given an explicit population, information
about the risk and a counterfactual distribution.
}
\details{
If \code{df} is missing and the original population (used for risk
estimation) can be accessed in \code{risk}, then the PAF will be calculated
from the original population.

In this function, the counterfactual population is defined by modifying
entries in the population data frame. This is done through the user-specified
function \code{mod_df}, which should accept a single input data frame and
outputs a similar data frame. The output data frame should have the same
column structure and number of rows as the original data frame.
}
\seealso{
\code{\link{paf_ext}} for PAF calculations with external prevalences
}
